removes web footer, adds list to main content, creates mobile personas component
