+++ components/PrelaunchHomePage/PrelaunchContent/PrelaunchContent.tsx
@@ -22,7 +22,14 @@ const PrelaunchContent = () => {
         <PrelaunchCta />
       </Flex>
       <Spacer />
-      <Flex justifyContent="flex-end" w="100%" h="auto" pb="10" pt="10" mt="20">
+      <Flex
+        justifyContent="flex-end"
+        w="100%"
+        h="500px"
+        pb="10"
+        pt="20"
+        mt="20"
+      >
         <DialectVideo />
       </Flex>
     </Flex>
