+++ components/PrelaunchHomePage/PrelaunchContent/PrelaunchMain.tsx
@@ -1,6 +1,17 @@
-import { Card, Flex, Heading, Spacer, Text } from '@chakra-ui/react'
+import {
+  Box,
+  Card,
+  Flex,
+  Heading,
+  ListItem,
+  SimpleGrid,
+  Spacer,
+  Text,
+  UnorderedList,
+} from '@chakra-ui/react'
 import React from 'react'
 import useMobileCheck from '../../hooks/useMobileCheck'
+import { personas } from './utils'
 
 const PrelaunchMain = () => {
   const isMobile = useMobileCheck()
@@ -13,18 +24,30 @@ const PrelaunchMain = () => {
     )
   }
   return (
-    <Flex w="575px" h="375px" flexDirection="column">
+    <Flex w="575px" h="675px" flexDirection="column">
       <Heading fontSize="4xl" fontWeight="black" lineHeight="1.25">
         Welcome to the future home of{' '}
         <Text color="brand.purple">ActingAccents.com!</Text>
       </Heading>
-      <Text fontSize="lg" mt="7" w="85%">
+      <Text fontSize="md" mt="7" w="85%">
         This website will be a resource for learning various dialects and
         accents for performance purposes. Whether you have a natural gift for
         dialects, or you find the process intimidating, having a systematic
         approach will help you to make more authentic dialect and accent
         choices.
       </Text>
+      <Box>
+        <Heading fontSize="xl">
+          Who Would Benefit From ActingAccents.com Courses?
+        </Heading>
+        <SimpleGrid columns={2}>
+          {personas.map((persona, index) => (
+            <UnorderedList key={index} pt={5} fontWeight="bold">
+              <ListItem>{persona}</ListItem>
+            </UnorderedList>
+          ))}
+        </SimpleGrid>
+      </Box>
       <Text fontSize="lg" fontWeight="extrabold" mt="5" w="85%">
         Ultimately, this course will teach you to teach yourself any dialect or
         accent you desire.
