+++ components/PrelaunchHomePage/PrelaunchContent/utils.tsx
@@ -1,0 +1,25 @@
+export const personas = [
+  'Stage, Film & Television Actors',
+  'Language Enthusiasts',
+  'Voiceover Artists',
+  'Directors',
+  'Dramaturgs',
+  'Secondary Education Teachers',
+  'Dialect and Accent Coaches',
+  'Singers',
+  'ESL Speakers Seeking Accent Modification',
+  'Writers',
+]
+
+export const personasMobile = [
+  'Stage, Film & Television Actors',
+  'Secondary Education Teachers',
+  'Dialect and Accent Coaches',
+  'ESL Speakers Seeking Accent Modification',
+  'Language Enthusiasts',
+  'Voiceover Artists',
+  'Directors',
+  'Dramaturgs',
+  'Singers',
+  'Writers',
+]
